+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## UNRELEASED
+- \[enhancement\] Correctly handle `encryptStream.destroy(error)` and `decryptStream.destroy(error)` functions.
+- \[deprecation\] Deprecate `encryptStream.emit('cancel')'` and `decryptStream.emit('cancel')` in favor of `stream.destroy()`
+
 ## v1.3.0 - 2025/03/20
 - \[new feature\] Expose keySize for AsymKeys
 - \[enhancement\] Update dependencies